server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
xgboost==2.0.3
python-dotenv==1.0.1

